requirements.txt
numpy==1.9.2
futures
matplotlib==1.3.1
pandas==0.18.1
scipy==0.9.0
sklearn==0.17.1
dlib
PySceneDetect==0.3.4
grpcio
PIL
cloghandler
youtube_dl
